+++ pkg/providers/cps/resource_akamai_cps_dv_enrollment_test.go
@@ -134,7 +134,7 @@ func TestResourceDVEnrollment(t *testing.T) {
 			},
 		}, nil).Once()
 
-		// final verification loop, everything in place
+		// final verification loop
 		client.On("GetChangeStatus", mock.Anything, cps.GetChangeStatusRequest{
 			EnrollmentID: 1,
 			ChangeID:     2,
@@ -712,7 +712,7 @@ func TestResourceDVEnrollment(t *testing.T) {
 			},
 		}, nil).Once()
 
-		// final verification loop, everything in place
+		// final verification loop
 		client.On("GetChangeStatus", mock.Anything, cps.GetChangeStatusRequest{
 			EnrollmentID: 1,
 			ChangeID:     2,
@@ -925,7 +925,7 @@ func TestResourceDVEnrollment(t *testing.T) {
 			},
 		}, nil).Once()
 
-		// final verification loop, everything in place
+		// final verification loop
 		client.On("GetChangeStatus", mock.Anything, cps.GetChangeStatusRequest{
 			EnrollmentID: 1,
 			ChangeID:     3,
@@ -1110,7 +1110,7 @@ func TestResourceDVEnrollment(t *testing.T) {
 			},
 		}, nil).Once()
 
-		// final verification loop, everything in place
+		// final verification loop
 		client.On("GetChangeStatus", mock.Anything, cps.GetChangeStatusRequest{
 			EnrollmentID: 1,
 			ChangeID:     2,
@@ -1867,7 +1867,7 @@ func TestResourceDVEnrollment(t *testing.T) {
 			},
 		}, nil).Once()
 
-		// final verification loop, everything in place
+		// final verification loop
 		client.On("GetChangeStatus", mock.Anything, cps.GetChangeStatusRequest{
 			EnrollmentID: 1,
 			ChangeID:     2,