+++ pkg/providers/property/resource_akamai_property.go
@@ -666,7 +666,6 @@ func resourcePropertyUpdate(ctx context.Context, d *schema.ResourceData, m inter
 	diags := diag.Diagnostics{}
 
 	immutable := []string{
-		"group_id",
 		"contract_id",
 		"product_id",
 		"property_id",
@@ -684,8 +683,9 @@ func resourcePropertyUpdate(ctx context.Context, d *schema.ResourceData, m inter
 	}
 
 	// We only update if these attributes change.
-	if !d.HasChanges("hostnames", "rules", "rule_format") {
-		logger.Debug("No changes to hostnames, rules, or rule_format (no update required)")
+	if !d.HasChanges("group_id", "hostnames", "rules", "rule_format") {
+		logger.Debug(
+			"No changes to group_id, hostnames, rules, or rule_format (no update required)")
 		return nil
 	}
 
@@ -713,7 +713,39 @@ func resourcePropertyUpdate(ctx context.Context, d *schema.ResourceData, m inter
 
 	propertyID := d.Id()
 	contractID := d.Get("contract_id").(string)
-	groupID := d.Get("group_id").(string)
+	oldGID, _ := d.GetChange("group_id")
+	oldGroupID := oldGID.(string)
+
+	// We want to change group id first: if the user loses access to the property as a result of
+	// group change, we want it to be visible immediately as an error during the same update.
+	//
+	// This way we will avoid the following scenario:
+	// 1. User changes the group and something else, e.g. hostnames and gets a success,
+	//    although they don't belong to the new group,
+	// 2. User changes hostnames again (not the group) and only then gets a (hard to understand)
+	//    error.
+	groupsDiffer, err := areGroupIDsDifferent(oldGroupID, property.GroupID)
+	if err != nil {
+		return diag.FromErr(err)
+	}
+	if groupsDiffer {
+		key := papiKey{
+			propertyID: property.PropertyID,
+			groupID:    oldGroupID,
+			contractID: property.ContractID,
+		}
+
+		err := updateGroupID(ctx, client, IAMClient(meta.Must(m)), key, property.GroupID)
+
+		if err != nil {
+			return diag.FromErr(err)
+		}
+
+		if !d.HasChanges("hostnames", "rules", "rule_format") {
+			logger.Debug("Only group_id changed, exiting early")
+			return resourcePropertyRead(ctx, d, m)
+		}
+	}
 
 	var propertyVersion int
 	if v, ok := d.GetOk("read_version"); ok && v.(int) != 0 {
@@ -722,7 +754,7 @@ func resourcePropertyUpdate(ctx context.Context, d *schema.ResourceData, m inter
 		propertyVersion = property.LatestVersion
 	}
 
-	resp, err := fetchPropertyVersion(ctx, client, propertyID, groupID, contractID, propertyVersion)
+	resp, err := fetchPropertyVersion(ctx, client, propertyID, property.GroupID, contractID, propertyVersion)
 	if err != nil {
 		d.Partial(true)
 		return diag.FromErr(err)
@@ -757,37 +789,44 @@ func resourcePropertyUpdate(ctx context.Context, d *schema.ResourceData, m inter
 		}
 	}
 
-	if !shouldUpdateRuleTree(d) {
-		return resourcePropertyRead(ctx, d, m)
+	if shouldUpdateRuleTree(d) {
+		if err := updateRuleTree(ctx, client, property, d); err != nil {
+			return diag.FromErr(err)
+		}
 	}
 
+	return resourcePropertyRead(ctx, d, m)
+}
+
+func updateRuleTree(ctx context.Context, client papi.PAPI, property papi.Property,
+	d *schema.ResourceData) error {
 	ruleFormat, err := tf.GetStringValue("rule_format", d)
 	if err != nil && !errors.Is(err, tf.ErrNotFound) {
-		return diag.FromErr(err)
+		return err
 	}
 
 	rulesJSON, err := tf.GetStringValue("rules", d)
 	if err != nil && !errors.Is(err, tf.ErrNotFound) {
-		return diag.FromErr(err)
+		return err
 	}
 
 	versionNotes, err := tf.GetStringValue("version_notes", tf.NewRawConfig(d))
 	if err != nil && !errors.Is(err, tf.ErrNotFound) {
-		return diag.FromErr(err)
+		return err
 	}
 
 	rulesUpdate, err := newRulesUpdate(rulesJSON, versionNotes)
 	if err != nil {
 		d.Partial(true)
-		return diag.FromErr(err)
+		return err
 	}
 
 	if err := updatePropertyRules(ctx, client, property, rulesUpdate, ruleFormat); err != nil {
 		d.Partial(true)
-		return diag.FromErr(err)
+		return err
 	}
 
-	return resourcePropertyRead(ctx, d, m)
+	return nil
 }
 
 func resourcePropertyDelete(ctx context.Context, d *schema.ResourceData, m interface{}) diag.Diagnostics {
@@ -805,7 +844,13 @@ func resourcePropertyDelete(ctx context.Context, d *schema.ResourceData, m inter
 	}
 	propertyID = d.Id()
 	contractID := str.AddPrefix(d.Get("contract_id").(string), "ctr_")
-	groupID := str.AddPrefix(d.Get("group_id").(string), "grp_")
+
+	// If delete is a result of a name update, group id could have also been changed.
+	// To be safe, we need to take the "old" value.
+	oldGroupID, newGroupID := d.GetChange("group_id")
+	logger.Debugf("resourcePropertyDelete: old group id: %s, new group id: %s",
+		oldGroupID.(string), newGroupID.(string))
+	groupID := str.AddPrefix(oldGroupID.(string), "grp_")
 
 	if err := removeProperty(ctx, client, propertyID, groupID, contractID); err != nil {
 		return diag.FromErr(err)