+++ pkg/providers/property/resource_akamai_property_bootstrap_test.go
@@ -6,6 +6,7 @@ import (
 	"strings"
 	"testing"
 
+	"github.com/akamai/AkamaiOPEN-edgegrid-golang/v8/pkg/iam"
 	"github.com/akamai/AkamaiOPEN-edgegrid-golang/v8/pkg/papi"
 	"github.com/akamai/terraform-provider-akamai/v6/pkg/common/str"
 	"github.com/akamai/terraform-provider-akamai/v6/pkg/common/testutils"
@@ -139,16 +140,17 @@ func TestBootstrapResourceUpdate(t *testing.T) {
 	tests := map[string]struct {
 		configPathForCreate string
 		configPathForUpdate string
-		init                func(*testing.T, *papi.Mock, testDataForPropertyBootstrap)
+		init                func(*testing.T, *papi.Mock, *iam.Mock, testDataForPropertyBootstrap)
 		mockData            testDataForPropertyBootstrap
 		errorForCreate      *regexp.Regexp
 		errorForUpdate      *regexp.Regexp
+		updateChecks        resource.TestCheckFunc
 	}{
 		"create and remove prefixes - no diff": {
 			configPathForCreate: "testdata/TestResPropertyBootstrap/create.tf",
 			configPathForUpdate: "testdata/TestResPropertyBootstrap/create_without_prefixes.tf",
-			init: func(t *testing.T, m *papi.Mock, data testDataForPropertyBootstrap) {
-				ExpectCreateProperty(m, data.name, data.groupID, data.contractID, data.productID, data.propertyID)
+			init: func(t *testing.T, papiMock *papi.Mock, iamMock *iam.Mock, data testDataForPropertyBootstrap) {
+				ExpectCreateProperty(papiMock, data.name, data.groupID, data.contractID, data.productID, data.propertyID)
 				prp := &papi.Property{
 					ContractID:   "ctr_2",
 					GroupID:      "grp_1",
@@ -156,8 +158,8 @@ func TestBootstrapResourceUpdate(t *testing.T) {
 					PropertyID:   "prp_123",
 					PropertyName: "property_name",
 				}
-				ExpectGetProperty(m, data.propertyID, data.groupID, data.contractID, prp)
-				ExpectRemoveProperty(m, data.propertyID, data.contractID, data.groupID)
+				ExpectGetProperty(papiMock, data.propertyID, data.groupID, data.contractID, prp)
+				ExpectRemoveProperty(papiMock, data.propertyID, data.contractID, data.groupID)
 			},
 			mockData: testDataForPropertyBootstrap{
 				propertyID: "prp_123",
@@ -166,21 +168,34 @@ func TestBootstrapResourceUpdate(t *testing.T) {
 				contractID: "ctr_2",
 				productID:  "prd_3",
 			},
+			updateChecks: checkPropertyBootstrapAttributes(testDataForPropertyBootstrap{
+				propertyID: "prp_123",
+				name:       "property_name",
+				groupID:    "grp_1",
+				contractID: "ctr_2",
+				productID:  "prd_3",
+			}),
 		},
-		"create and update group - error": {
+		"create and update group id": {
 			configPathForCreate: "testdata/TestResPropertyBootstrap/create.tf",
 			configPathForUpdate: "testdata/TestResPropertyBootstrap/update_group.tf",
-			init: func(t *testing.T, m *papi.Mock, data testDataForPropertyBootstrap) {
-				ExpectCreateProperty(m, data.name, data.groupID, data.contractID, data.productID, data.propertyID)
+			init: func(t *testing.T, papiMock *papi.Mock, iamMock *iam.Mock, data testDataForPropertyBootstrap) {
+				ExpectCreateProperty(papiMock, data.name, data.groupID, data.contractID, data.productID,
+					data.propertyID).Once()
 				prp := &papi.Property{
+					AssetID:      "aid_55555",
 					ContractID:   "ctr_2",
 					GroupID:      "grp_1",
 					ProductID:    "prd_3",
 					PropertyID:   "prp_123",
 					PropertyName: "property_name",
 				}
-				ExpectGetProperty(m, data.propertyID, data.groupID, data.contractID, prp)
-				ExpectRemoveProperty(m, data.propertyID, data.contractID, data.groupID)
+				ExpectGetProperty(papiMock, data.propertyID, data.groupID, data.contractID, prp).Times(3)
+
+				mockMoveProperty(iamMock, 55555, 1, 111)
+
+				ExpectGetProperty(papiMock, data.propertyID, "grp_111", data.contractID, prp).Twice()
+				ExpectRemoveProperty(papiMock, data.propertyID, data.contractID, "grp_111")
 			},
 			mockData: testDataForPropertyBootstrap{
 				propertyID: "prp_123",
@@ -189,13 +204,101 @@ func TestBootstrapResourceUpdate(t *testing.T) {
 				contractID: "ctr_2",
 				productID:  "prd_3",
 			},
-			errorForUpdate: regexp.MustCompile("updating field `group_id` is not possible"),
+			updateChecks: checkPropertyBootstrapAttributes(testDataForPropertyBootstrap{
+				propertyID: "prp_123",
+				name:       "property_name",
+				groupID:    "grp_111",
+				contractID: "ctr_2",
+				productID:  "prd_3",
+			}),
+		},
+		"create and update name - resource replacement": {
+			configPathForCreate: "testdata/TestResPropertyBootstrap/create.tf",
+			configPathForUpdate: "testdata/TestResPropertyBootstrap/update_name.tf",
+			init: func(t *testing.T, papiMock *papi.Mock, iamMock *iam.Mock, data testDataForPropertyBootstrap) {
+				ExpectCreateProperty(papiMock, data.name, data.groupID, data.contractID, data.productID,
+					data.propertyID).Once()
+				prp := &papi.Property{
+					AssetID:      "aid_55555",
+					ContractID:   "ctr_2",
+					GroupID:      "grp_1",
+					ProductID:    "prd_3",
+					PropertyID:   "prp_123",
+					PropertyName: "property_name",
+				}
+				ExpectGetProperty(papiMock, data.propertyID, data.groupID, data.contractID, prp).Times(2)
+
+				ExpectRemoveProperty(papiMock, data.propertyID, data.contractID, data.groupID).Once()
+
+				ExpectCreateProperty(papiMock, "property_name2", data.groupID, data.contractID, data.productID,
+					data.propertyID).Once()
+
+				ExpectGetProperty(papiMock, data.propertyID, data.groupID, data.contractID, prp).Once()
+
+				ExpectRemoveProperty(papiMock, data.propertyID, data.contractID, data.groupID).Once()
+
+			},
+			mockData: testDataForPropertyBootstrap{
+				propertyID: "prp_123",
+				name:       "property_name",
+				groupID:    "grp_1",
+				contractID: "ctr_2",
+				productID:  "prd_3",
+			},
+			updateChecks: checkPropertyBootstrapAttributes(testDataForPropertyBootstrap{
+				propertyID: "prp_123",
+				name:       "property_name2",
+				groupID:    "grp_1",
+				contractID: "ctr_2",
+				productID:  "prd_3",
+			}),
+		},
+		"create and update name and group id - resource replacement": {
+			configPathForCreate: "testdata/TestResPropertyBootstrap/create.tf",
+			configPathForUpdate: "testdata/TestResPropertyBootstrap/update_name_and_group.tf",
+			init: func(t *testing.T, papiMock *papi.Mock, iamMock *iam.Mock, data testDataForPropertyBootstrap) {
+				ExpectCreateProperty(papiMock, data.name, data.groupID, data.contractID, data.productID,
+					data.propertyID).Once()
+				prp := &papi.Property{
+					AssetID:      "aid_55555",
+					ContractID:   "ctr_2",
+					GroupID:      "grp_1",
+					ProductID:    "prd_3",
+					PropertyID:   "prp_123",
+					PropertyName: "property_name",
+				}
+				ExpectGetProperty(papiMock, data.propertyID, data.groupID, data.contractID, prp).Times(2)
+
+				ExpectRemoveProperty(papiMock, data.propertyID, data.contractID, data.groupID).Once()
+
+				ExpectCreateProperty(papiMock, "property_name2", "grp_93", data.contractID, data.productID,
+					data.propertyID).Once()
+
+				ExpectGetProperty(papiMock, data.propertyID, "grp_93", data.contractID, prp).Once()
+
+				ExpectRemoveProperty(papiMock, data.propertyID, data.contractID, "grp_93").Once()
+
+			},
+			mockData: testDataForPropertyBootstrap{
+				propertyID: "prp_123",
+				name:       "property_name",
+				groupID:    "grp_1",
+				contractID: "ctr_2",
+				productID:  "prd_3",
+			},
+			updateChecks: checkPropertyBootstrapAttributes(testDataForPropertyBootstrap{
+				propertyID: "prp_123",
+				name:       "property_name2",
+				groupID:    "grp_93",
+				contractID: "ctr_2",
+				productID:  "prd_3",
+			}),
 		},
 		"create and update contract - error": {
 			configPathForCreate: "testdata/TestResPropertyBootstrap/create.tf",
 			configPathForUpdate: "testdata/TestResPropertyBootstrap/update_contract.tf",
-			init: func(t *testing.T, m *papi.Mock, data testDataForPropertyBootstrap) {
-				ExpectCreateProperty(m, data.name, data.groupID, data.contractID, data.productID, data.propertyID)
+			init: func(t *testing.T, papiMock *papi.Mock, iamMock *iam.Mock, data testDataForPropertyBootstrap) {
+				ExpectCreateProperty(papiMock, data.name, data.groupID, data.contractID, data.productID, data.propertyID)
 				prp := &papi.Property{
 					ContractID:   "ctr_2",
 					GroupID:      "grp_1",
@@ -203,8 +306,8 @@ func TestBootstrapResourceUpdate(t *testing.T) {
 					PropertyID:   "prp_123",
 					PropertyName: "property_name",
 				}
-				ExpectGetProperty(m, data.propertyID, data.groupID, data.contractID, prp)
-				ExpectRemoveProperty(m, data.propertyID, data.contractID, data.groupID)
+				ExpectGetProperty(papiMock, data.propertyID, data.groupID, data.contractID, prp)
+				ExpectRemoveProperty(papiMock, data.propertyID, data.contractID, data.groupID)
 			},
 			mockData: testDataForPropertyBootstrap{
 				propertyID: "prp_123",
@@ -218,8 +321,8 @@ func TestBootstrapResourceUpdate(t *testing.T) {
 		"create and update product - error": {
 			configPathForCreate: "testdata/TestResPropertyBootstrap/create.tf",
 			configPathForUpdate: "testdata/TestResPropertyBootstrap/update_product.tf",
-			init: func(t *testing.T, m *papi.Mock, data testDataForPropertyBootstrap) {
-				ExpectCreateProperty(m, data.name, data.groupID, data.contractID, data.productID, data.propertyID)
+			init: func(t *testing.T, papiMock *papi.Mock, iamMock *iam.Mock, data testDataForPropertyBootstrap) {
+				ExpectCreateProperty(papiMock, data.name, data.groupID, data.contractID, data.productID, data.propertyID)
 				prp := &papi.Property{
 					ContractID:   "ctr_2",
 					GroupID:      "grp_1",
@@ -227,8 +330,8 @@ func TestBootstrapResourceUpdate(t *testing.T) {
 					PropertyID:   "prp_123",
 					PropertyName: "property_name",
 				}
-				ExpectGetProperty(m, data.propertyID, data.groupID, data.contractID, prp)
-				ExpectRemoveProperty(m, data.propertyID, data.contractID, data.groupID)
+				ExpectGetProperty(papiMock, data.propertyID, data.groupID, data.contractID, prp)
+				ExpectRemoveProperty(papiMock, data.propertyID, data.contractID, data.groupID)
 			},
 			mockData: testDataForPropertyBootstrap{
 				propertyID: "prp_123",
@@ -246,32 +349,34 @@ func TestBootstrapResourceUpdate(t *testing.T) {
 		t.Run(name, func(t *testing.T) {
 			t.Parallel()
 
-			m := &papi.Mock{}
+			papiMock := &papi.Mock{}
+			iamMock := &iam.Mock{}
 			if test.init != nil {
-				test.init(t, m, test.mockData)
+				test.init(t, papiMock, iamMock, test.mockData)
 			}
 
-			useClient(m, nil, func() {
-				resource.UnitTest(t, resource.TestCase{
-					ProtoV6ProviderFactories: testutils.NewProtoV6ProviderFactory(NewSubprovider()),
-					IsUnitTest:               true,
-					Steps: []resource.TestStep{
-						{
-							Config:      testutils.LoadFixtureString(t, test.configPathForCreate),
-							Check:       checkPropertyBootstrapAttributes(test.mockData),
-							ExpectError: test.errorForCreate,
+			useClient(papiMock, nil, func() {
+				useIam(iamMock, func() {
+					resource.UnitTest(t, resource.TestCase{
+						ProtoV6ProviderFactories: testutils.NewProtoV6ProviderFactory(NewSubprovider()),
+						IsUnitTest:               true,
+						Steps: []resource.TestStep{
+							{
+								Config:      testutils.LoadFixtureString(t, test.configPathForCreate),
+								Check:       checkPropertyBootstrapAttributes(test.mockData),
+								ExpectError: test.errorForCreate,
+							},
+							{
+								Config:      testutils.LoadFixtureString(t, test.configPathForUpdate),
+								Check:       test.updateChecks,
+								ExpectError: test.errorForUpdate,
+							},
 						},
-						{
-							Config:             testutils.LoadFixtureString(t, test.configPathForUpdate),
-							PlanOnly:           true,
-							ExpectNonEmptyPlan: false,
-							ExpectError:        test.errorForUpdate,
-						},
-					},
+					})
 				})
 			})
 
-			m.AssertExpectations(t)
+			papiMock.AssertExpectations(t)
 		})
 	}
 }