+++ pkg/providers/property/testdata/TestGroupIDUpdate/base.tf
@@ -1,0 +1,16 @@
+provider "akamai" {
+  edgerc = "../../common/testutils/edgerc"
+}
+
+resource "akamai_property" "test" {
+  name        = "dummy_name"
+  group_id    = "grp_1"
+  contract_id = "ctr_2"
+  product_id  = "prd_3"
+
+  hostnames {
+    cname_to               = "to.test.domain"
+    cname_from             = "from.test.domain"
+    cert_provisioning_type = "DEFAULT"
+  }
+}